+++ src/components/index.ts
@@ -181,3 +181,5 @@ export * from './Timeline'
 export * from './Treeview'
 
 export * from './QRCode'
+
+export * from './Watermark'